The LATokenFieldCell will (optionally) animate to/from scaled mode.
After long press 'LATokenField' will do hit test on subviews. If a 'LATokenFieldCell' passes the hit test, it becomes scaled to 120%, semi-transparent and draggable. It can be moved to a new index between other cells. When the long press is released, it is set to the new index and becomes normal (unscaled and unselected).

If a token field cell is selected (or draggable) the caret is hidden.