Delegates of LATokenField can provide a subclass of LATokenFieldCell for a given represented object.
The goal is to provide a LATokenFieldCell subclass that 'knows' the represented object and is capable adapting the content and style of each cell according to the represented object state.

There is a new example called 'Colors'. It changes the background of the token field cell based on a UIColor represented object.

Another solution is using a decorator, instead of subclassing LATokenFieldCell. This is a work in progress.